// apps/api/src/common/exceptions/exception-codes.ts
export enum ExceptionCode {
  INTERNAL_ERROR = 'INTERNAL_ERROR',

  // auth
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',

  // tenant
  TENANT_NOT_FOUND = 'TENANT_NOT_FOUND',
  TENANT_ALREADY_EXISTS = 'TENANT_ALREADY_EXISTS',

  // building
  BUILDING_NOT_FOUND = 'BUILDING_NOT_FOUND',

  // lease
  LEASE_NOT_FOUND = 'LEASE_NOT_FOUND',

  // generic
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  DUPLICATE_RESOURCE = 'DUPLICATE_RESOURCE',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
}